switch from own build connection handler and threadpool to `tokio`
We currently do not read the rest of the request body, if it doesn't fit
into the last buffer. Reperating socket reads into "headers" and "rest-of-
body" wasn't a good idea anyway. It complicated things more then needed,
which is good illustrated by the amount of removed lines from
`handle_connection()`. Re-adding the read of the body should be rather
easy now.

The ease-of-use of `tokio` really supprised me, good stuff!